STM32Cube_FW_F4_V1.23.0 Released 2018-11-23
STM32CubeF4 Firmware Package V1.23.0 / November 23, 2018

Main Changes
* Add support to TouchGFX stack
* Update version of STM32CubeF4GettingStarted.pdf file.

=====================================================================

Note: Git Large File Support (LFS) needed for the file
      "Middlewares/ST/TouchGFX/installer/TouchGFX-4.10.0.msi"

+++ Projects/STM32F412G-Discovery/Applications/FreeRTOS/FreeRTOS_MPU/Inc/FreeRTOSConfig.h
@@ -193,7 +193,7 @@ standard names. */
 
 /* Normal assert() semantics without relying on the provision of an assert.h
 header file. */
-#define configASSERT( x ) if( ( x ) == 0UL ) { taskDISABLE_INTERRUPTS(); for( ;; ); }
+#define configASSERT( x ) if((uint32_t)( x ) == 0UL ) { taskDISABLE_INTERRUPTS(); for( ;; ); }
 
 #ifdef __cplusplus
 }
